+++ activerecord/lib/active_record/persistence.rb
@@ -708,6 +708,7 @@ module ActiveRecord
     def create_or_update(*args, &block)
       _raise_readonly_record_error if readonly?
       return false if destroyed?
+      self.class.connection.add_transaction_saved_record(self)
       result = new_record? ? _create_record(&block) : _update_record(*args, &block)
       result != false
     end